pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "invoice-collector"
version = "0.1.0"
description = "自动从QQ邮箱下载发票并按月归档"
requires-python = ">=3.10"
dependencies = [
    "pyyaml",
    "playwright",
    "pdfplumber",
    "pypdf",
    "click",
    "rich",
    "python-dateutil",
]

[project.scripts]
agentinvoice = "invoice_collector.main:main"

[tool.setuptools.packages.find]
where = ["src"]
